--- Behavioral.Automation.Bindings.UI/Abstractions/ICheckboxElement.cs
@@ -1,6 +1,0 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
-
-public interface ICheckboxElement : IWebElement
-{
-    public Task ClickAsync();
-}
