--- Behavioral.Automation.Bindings.UI/Abstractions/ITableWrapper.cs
@@ -1,5 +1,0 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
-
-public interface ITableWrapper : IWebElement
-{
-}
