+++ Behavioral.Automation.Interfaces.UI/Abstractions/ElementSelector.cs
@@ -1,4 +1,4 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
+﻿namespace Behavioral.Automation.Interfaces.UI.Abstractions;
 
 public class ElementSelector
 {
