+++ Behavioral.Automation.Interfaces.UI/Abstractions/IPage.cs
@@ -1,10 +1,9 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
+﻿namespace Behavioral.Automation.Interfaces.UI.Abstractions;
 
 public interface IPage
 {
     public Task GoToUrlAsync(string url);
     public Task GoToApplicationUrlAsync();
     public Task HaveTitleAsync(string title);
-    public Task<IFileChooser> RunAndWaitForFileChooserAsync(Func<Task> action);
     public Task ScreenshotAsync(string path);
 }
