+++ Behavioral.Automation.Interfaces.UI/Abstractions/IWebElementStorageService.cs
@@ -1,4 +1,4 @@
-﻿namespace Behavioral.Automation.Bindings.UI.Abstractions;
+﻿namespace Behavioral.Automation.Interfaces.UI.Abstractions;
 
 public interface IWebElementStorageService
 {
